@@ -5,25 +5,28 @@ using Npgsql;
 using Microsoft.Extensions.Logging;
 namespace DBHelper
 {
-	public abstract class PgExecute : DBConnection
+	public abstract class PgExecute
 	{
-		public static ILogger _logger = null;
-		private NpgsqlTransaction _transaction = null;
-		public PgExecute(ILogger logger)
+		public static ILogger _logger;
+		public NpgsqlTransaction _transaction;
+		ConnectionPool _pool;
+
+		protected PgExecute(int poolSize, string connectionString, ILogger logger)
 		{
 			_logger = logger;
+			_pool = new ConnectionPool(poolSize, connectionString);
 		}
-		public PgExecute() { }
+
 
 		/// <summary>
 		/// 执行命令前准备
 		/// </summary>
 		protected void PrepareCommand(NpgsqlCommand command, CommandType commandType, string commandText, NpgsqlParameter[] commandParameters)
 		{
-			if (commandText == null || commandText.Length == 0 || command == null) throw new ArgumentNullException("commandText error");
-			if (_connection == null)
-				_connection = GetConnection();
-			command.Connection = _connection;
+			if (_pool == null) throw new ArgumentException("Connection Pool is null");
+			if (commandText.IsNullOrEmpty() || command == null) throw new ArgumentNullException("Command is error");
+			var conn = _pool.GetConnection();
+			command.Connection = conn;
 			command.CommandText = commandText;
 			command.CommandType = commandType;
 			if (commandParameters != null)
@@ -47,7 +50,6 @@ namespace DBHelper
 			try
 			{
 				PrepareCommand(cmd, commandType, commandText, commandParameters);
-				OpenConnection(cmd.Connection);
 				ret = cmd.ExecuteScalar();
 
 			}
@@ -58,8 +60,7 @@ namespace DBHelper
 			}
 			finally
 			{
-				//mark: 有事务不能直接释放命令
-				if (_transaction != null)
+				if (_transaction == null || _transaction.IsCompleted)
 					Close(cmd, cmd.Connection);
 			}
 			return ret;
@@ -74,7 +75,6 @@ namespace DBHelper
 			try
 			{
 				PrepareCommand(cmd, commandType, commandText, commandParameters);
-				OpenConnection(cmd.Connection);
 				ret = cmd.ExecuteNonQuery();
 			}
 			catch (Exception ex)
@@ -84,8 +84,7 @@ namespace DBHelper
 			}
 			finally
 			{
-				//mark: 若有事务不能直接释放命令
-				if (_transaction != null)
+				if (_transaction == null || _transaction.IsCompleted)
 					Close(cmd, cmd.Connection);
 			}
 			return ret;
@@ -99,8 +98,6 @@ namespace DBHelper
 			try
 			{
 				PrepareCommand(cmd, commandType, commandText, commandParameters);
-				if (cmd.Connection.State != ConnectionState.Open)
-					cmd.Connection.Open();
 				using (reader = cmd.ExecuteReader())
 					while (reader.Read())
 						action?.Invoke(reader);
@@ -112,9 +109,9 @@ namespace DBHelper
 			}
 			finally
 			{
-				if (_transaction == null)
+				if (_transaction == null || _transaction.IsCompleted)
 					Close(cmd, cmd.Connection);
-				if (!reader.IsClosed)
+				if (reader != null && !reader.IsClosed)
 					reader.Close();
 			}
 		}
@@ -125,13 +122,11 @@ namespace DBHelper
 		{
 			string str = string.Empty;
 			if (cmd.Parameters != null)
-			{
 				foreach (NpgsqlParameter item in cmd.Parameters)
 					str += $"{item.ParameterName}:{item.Value}\n";
-			}
+			if (_transaction != null)
+				RollBackTransaction();
 			Close(cmd, cmd.Connection);
-			//mark: 如果有事务, 则回滚事务
-			RollBackTransaction();
 			//done: 输出错误日志
 			_logger.LogError(new EventId(111111), ex, "数据库执行出错：===== \n {0}\n{1}\n{2}", cmd.CommandText, cmd.Parameters, str);//输出日志
 
@@ -147,11 +142,9 @@ namespace DBHelper
 					cmd.Parameters.Clear();
 				cmd.Dispose();
 			}
-			if (connection != null)
-			{
-				StopConnection(connection);
-				//mark: 释放连接委托
-			}
+			_pool.ReleaseConnection(connection);
+			if (_transaction != null)
+				_transaction.Dispose();
 		}
 		#region 事务
 		/// <summary>
@@ -159,14 +152,13 @@ namespace DBHelper
 		/// </summary>
 		public void BeginTransaction()
 		{
-			if (_transaction == null)
+			if (_transaction != null)
 				throw new Exception("the transaction is opened");
-			_connection = GetConnection();
-			OpenConnection(_connection);
-			_transaction = _connection.BeginTransaction();
+			var conn = _pool.GetConnection();
+			_transaction = conn.BeginTransaction();
 		}
 		/// <summary>
-		/// 确认是事务
+		/// 确认事务
 		/// </summary>
 		public void CommitTransaction()
 		{
