@@ -5,15 +5,23 @@ using System.Text;
 
 namespace DBHelper
 {
-	public class UnionModel
+	public class Union
 	{
 		public string AliasName { get; set; }
-		public Type Model { get; set; }
+		public string Table { get; set; }
 		public string Expression { get; set; }
-		public UnionType UnionType { get; set; }
+		public UnionEnum UnionType { get; set; }
+		public Union(string aliasName, string table, string expression, UnionEnum unionType)
+		{
+			AliasName = aliasName;
+			Table = table;
+			Expression = expression;
+			UnionType = unionType;
+		}
 	}
-	public enum UnionType
+	public enum UnionEnum
 	{
 		INNER_JOIN = 1, LEFT_JOIN, RIGHT_JOIN, LEFT_OUTER_JOIN, RIGHT_OUTER_JOIN
 	}
 }
+
