@@ -4,8 +4,26 @@ using System.Linq;
 using System.Threading.Tasks;
 using NpgsqlTypes;
 
-namespace DBHelper.CodeFactory
+namespace Common.CodeFactory
 {
+	public class GenerateModel
+	{
+		/// <summary>
+		/// 数据库连接字符串
+		/// </summary>
+		public string ConnectionString { get; set; }
+		/// <summary>
+		/// 项目名称
+		/// </summary>
+		public string ProjectName { get; set; }
+		/// <summary>
+		/// 输出路径
+		/// </summary>
+		public string OutputPath { get; set; }
+		/// <summary>
+		/// 数据库
+		/// </summary>
+	}
 	public class TableViewModel
 	{
 		public string Name { get; set; }
@@ -47,21 +65,58 @@ namespace DBHelper.CodeFactory
 		public string RefColumn { get; set; }
 		public string TableName { get; set; }
 		public string Nspname { get; set; }
+		public bool IsOneToOne { get; set; }
 	}
 	public class ConstraintMoreToMore
 	{
-		public string MainTable { get; set; }//主要输出表
-		public string MainNspname { get; set; }//主要输出表的schema
-		public string MainField { get; set; }//主要输出表的主键
-		public string MinorField { get; set; }//当前表的字段名称
-		public string MinorNspname { get; set; }//当前表的schema
-		public string MinorTable { get; set; }//当前表名
-		public string CenterNspname { get; set; }//中间表的schema
-		public string CenterMainField { get; set; }//中间表中主要输出表的主键字段名
-		public string CenterMainType { get; set; }//中间表中主要输出表的主键字段类型
-		public string CenterMinorField { get; set; }//中间表中当前表的主键字段名
-		public string CenterMinorType { get; set; }//中间表中当前表的主键字段类型
-		public string CenterTable { get; set; }//中间表的表名
+		/// <summary>
+		/// 主要输出表
+		/// </summary>
+		public string MainTable { get; set; }
+		/// <summary>
+		/// 主要输出表的Schema名称
+		/// </summary>
+		public string MainNspname { get; set; }
+		/// <summary>
+		/// 主要输出表的主键
+		/// </summary>
+		public string MainField { get; set; }
+		/// <summary>
+		/// 当前表的字段名称
+		/// </summary>
+		public string MinorField { get; set; }
+		/// <summary>
+		/// 当前表的Schema名称
+		/// </summary>
+		public string MinorNspname { get; set; }
+		/// <summary>
+		/// 当前表名称
+		/// </summary>
+		public string MinorTable { get; set; }
+		/// <summary>
+		/// 中间表的Schema
+		/// </summary>
+		public string CenterNspname { get; set; }
+		/// <summary>
+		/// 中间表中主要输出表的主键字段名称
+		/// </summary>
+		public string CenterMainField { get; set; }
+		/// <summary>
+		/// 中间表中主要输出表的主键字段类型
+		/// </summary>
+		public string CenterMainType { get; set; }
+		/// <summary>
+		/// 中间表中当前表的主键字段名称
+		/// </summary>
+		public string CenterMinorField { get; set; }
+		/// <summary>
+		/// 中间表中当前表的主键字段类型
+		/// </summary>
+		public string CenterMinorType { get; set; }
+		/// <summary>
+		/// 中间表的表名
+		/// </summary>
+		public string CenterTable { get; set; }
 	}
 	public class EnumTypeInfo
 	{
